# Conflicts:
#	app/src/main/java/com/example/myapplication/util/devicesvm/LampViewModel.kt

+++ app/src/main/java/com/example/myapplication/data/LampUiState.kt
@@ -19,5 +19,5 @@ data class LampType(
 data class LampState(
     val status: String = "off",
     val color: String = "FFFFFF",
-    val brightness: Double = 100.0
+    val brightness: Int = 100
 )
